+++ android/src/main/java/io/openim/flutter_openim_sdk/manager/MessageManager.java
@@ -417,6 +417,13 @@ public class MessageManager extends BaseManager {
                 jsonValue(methodCall)
         );
     }
+    public void getAdvancedHistoryMessageList(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.getAdvancedHistoryMessageList(
+        new OnBaseListener(result, methodCall),
+                value(methodCall, "operationID"),
+                jsonValue(methodCall)
+        );
+    }
 
     public void sendUploadMessage(MethodCall methodCall, MethodChannel.Result result) {
         Open_im_sdk.sendUploadMessage(
@@ -446,5 +453,72 @@ public class MessageManager extends BaseManager {
                 value(methodCall, "count"),
                 int2long(methodCall, "startTime")
         );
+    }
+    public void findMessageList(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.findMessageList(
+                new OnBaseListener(result, methodCall),
+                value(methodCall, "operationID"),
+                jsonValue(methodCall, "searchParams")
+        );
+    }
+
+    public void createAdvancedTextMessage(MethodCall methodCall, MethodChannel.Result result) {
+        CommonUtil.runMainThreadReturn(result,
+                Open_im_sdk.createAdvancedTextMessage(
+                        value(methodCall, "operationID"),
+                        value(methodCall, "text"),
+                        jsonValue(methodCall, "richMessageInfoList")
+                ));
+    }
+
+    public void createAdvancedQuoteMessage(MethodCall methodCall, MethodChannel.Result result) {
+        CommonUtil.runMainThreadReturn(result,
+                Open_im_sdk.createAdvancedQuoteMessage(
+                        value(methodCall, "operationID"),
+                        value(methodCall, "quoteText"),
+                        jsonValue(methodCall, "quoteMessage"),
+                        jsonValue(methodCall, "richMessageInfoList")
+                ));
+    }
+
+    public void sendMessageNotOss(MethodCall methodCall, MethodChannel.Result result) {
+        Open_im_sdk.sendMessageNotOss(
+                new OnMsgSendListener(result, methodCall),
+                value(methodCall, "operationID"),
+                jsonValue(methodCall, "message"),
+                value(methodCall, "userID"),
+                value(methodCall, "groupID"),
+                jsonValue(methodCall, "offlinePushInfo")
+        );
+    }
+
+    public void createImageMessageByURL(MethodCall methodCall, MethodChannel.Result result) {
+        CommonUtil.runMainThreadReturn(result,
+                Open_im_sdk.createImageMessageByURL(
+                        value(methodCall, "operationID"),
+                        jsonValue(methodCall, "sourcePicture"),
+                        jsonValue(methodCall, "bigPicture"),
+                        jsonValue(methodCall, "snapshotPicture")));
+    }
+
+    public void createSoundMessageByURL(MethodCall methodCall, MethodChannel.Result result) {
+        CommonUtil.runMainThreadReturn(result,
+                Open_im_sdk.createSoundMessageByURL(
+                        value(methodCall, "operationID"),
+                        jsonValue(methodCall, "soundElem")));
+    }
+
+    public void createVideoMessageByURL(MethodCall methodCall, MethodChannel.Result result) {
+        CommonUtil.runMainThreadReturn(result,
+                Open_im_sdk.createVideoMessageByURL(
+                        value(methodCall, "operationID"),
+                        jsonValue(methodCall, "videoElem")));
+    }
+
+    public void createFileMessageByURL(MethodCall methodCall, MethodChannel.Result result) {
+        CommonUtil.runMainThreadReturn(result,
+                Open_im_sdk.createFileMessageByURL(
+                        value(methodCall, "operationID"),
+                        jsonValue(methodCall, "fileElem")));
     }
 }