+++ lib/src/enum/message_type.dart
@@ -45,8 +45,18 @@ class MessageType {
   /// 自定义表情
   static const custom_face = 115;
 
+  /// 群消息已读回执
+  static const groupHasReadReceipt = 116;
+
+  /// 富文本消息
+  static const advancedText = 117;
+
   /// 高级撤回
   static const advancedRevoke = 118;
+
+  static const customMsgNotTriggerConversation = 119;
+
+  static const customMsgOnlineOnly = 120;
 
   /// 通知类型
   static const notificationBegin = 1000;