+++ lib/src/manager/im_manager.dart
@@ -248,31 +248,53 @@ class IMManager {
         } else if (call.method == ListenerType.signalingListener) {
           String type = call.arguments['type'];
           dynamic data = call.arguments['data'];
-          final u = Utils.toObj(data, (map) => SignalingInfo.fromJson(map));
+          dynamic info;
+          switch (type) {
+            case 'onRoomParticipantConnected':
+            case 'onRoomParticipantDisconnected':
+              info = Utils.toObj(data, (map) => RoomCallingInfo.fromJson(map));
+              break;
+            case 'onStreamChange':
+              info =
+                  Utils.toObj(data, (map) => MeetingStreamEvent.fromJson(map));
+              break;
+            default:
+              info = Utils.toObj(data, (map) => SignalingInfo.fromJson(map));
+              break;
+          }
           switch (type) {
             case 'onInvitationCancelled':
-              signalingManager.listener.invitationCancelled(u);
+              signalingManager.listener.invitationCancelled(info);
               break;
             case 'onInvitationTimeout':
-              signalingManager.listener.invitationTimeout(u);
+              signalingManager.listener.invitationTimeout(info);
               break;
             case 'onInviteeAccepted':
-              signalingManager.listener.inviteeAccepted(u);
+              signalingManager.listener.inviteeAccepted(info);
               break;
             case 'onInviteeRejected':
-              signalingManager.listener.inviteeRejected(u);
+              signalingManager.listener.inviteeRejected(info);
               break;
             case 'onReceiveNewInvitation':
-              signalingManager.listener.receiveNewInvitation(u);
+              signalingManager.listener.receiveNewInvitation(info);
               break;
             case 'onInviteeAcceptedByOtherDevice':
-              signalingManager.listener.inviteeAcceptedByOtherDevice(u);
+              signalingManager.listener.inviteeAcceptedByOtherDevice(info);
               break;
             case 'onInviteeRejectedByOtherDevice':
-              signalingManager.listener.inviteeRejectedByOtherDevice(u);
+              signalingManager.listener.inviteeRejectedByOtherDevice(info);
               break;
             case 'onHangUp':
-              signalingManager.listener.hangup(u);
+              signalingManager.listener.hangup(info);
+              break;
+            case 'onRoomParticipantConnected':
+              signalingManager.listener.roomParticipantConnected(info);
+              break;
+            case 'onRoomParticipantDisconnected':
+              signalingManager.listener.roomParticipantDisconnected(info);
+              break;
+            case 'onStreamChange':
+              signalingManager.listener.streamChangedEvent(info);
               break;
           }
         } else if (call.method == ListenerType.workMomentsListener) {
@@ -313,6 +335,7 @@ class IMManager {
     required OnConnectListener listener,
     int logLevel = 6,
     String objectStorage = 'cos',
+    String? encryptionKey,
     String? operationID,
   }) {
     this._connectListener = listener;
@@ -327,6 +350,7 @@ class IMManager {
             "data_dir": dataDir,
             "log_level": logLevel,
             "object_storage": objectStorage,
+            "encryption_key": encryptionKey,
             "operationID": Utils.checkOperationID(operationID),
           },
         ));
@@ -405,6 +429,19 @@ class IMManager {
             'path': path,
             'token': token ?? this.token,
             'obj': objectStorage ?? this._objectStorage,
+            'operationID': Utils.checkOperationID(operationID),
+          }));
+
+  /// 更新firebase客户端注册token
+  /// [fcmToken] firebase token
+  Future updateFcmToken({
+    required String fcmToken,
+    String? operationID,
+  }) =>
+      _channel.invokeMethod(
+          'updateFcmToken',
+          _buildParam({
+            'fcmToken': fcmToken,
             'operationID': Utils.checkOperationID(operationID),
           }));