+++ lib/src/models/group_info.dart
@@ -140,6 +140,16 @@ class GroupInfo {
   int get sessionType => groupType == GroupType.general
       ? ConversationType.group
       : ConversationType.superGroup;
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is GroupInfo &&
+          runtimeType == other.runtimeType &&
+          groupID == other.groupID;
+
+  @override
+  int get hashCode => groupID.hashCode;
 }
 
 /// 群成员信息
@@ -226,6 +236,17 @@ class GroupMembersInfo {
     data['inviterUserID'] = this.inviterUserID;
     return data;
   }
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is GroupMembersInfo &&
+          runtimeType == other.runtimeType &&
+          groupID == other.groupID &&
+          userID == other.userID;
+
+  @override
+  int get hashCode => groupID.hashCode ^ userID.hashCode;
 }
 
 /// 群成员角色
@@ -319,6 +340,12 @@ class GroupApplicationInfo {
   /// 扩展信息
   String? ex;
 
+  /// 2：通过邀请  3：通过搜索  4：通过二维码
+  int? joinSource;
+
+  /// 邀请进群用户ID
+  String? inviterUserID;
+
   GroupApplicationInfo({
     this.groupID,
     this.groupName,
@@ -342,6 +369,8 @@ class GroupApplicationInfo {
     this.handleUserID,
     this.handledTime,
     this.ex,
+    this.inviterUserID,
+    this.joinSource,
   });
 
   GroupApplicationInfo.fromJson(Map<String, dynamic> json) {
@@ -367,6 +396,8 @@ class GroupApplicationInfo {
     handleUserID = json['handleUserID'];
     handledTime = json['handledTime'];
     ex = json['ex'];
+    inviterUserID = json['inviterUserID'];
+    joinSource = json['joinSource'];
   }
 
   Map<String, dynamic> toJson() {
@@ -393,6 +424,8 @@ class GroupApplicationInfo {
     data['handleUserID'] = this.handleUserID;
     data['handledTime'] = this.handledTime;
     data['ex'] = this.ex;
+    data['inviterUserID'] = this.inviterUserID;
+    data['joinSource'] = this.joinSource;
     return data;
   }
 }